+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Changes
 
-`program::abort` has been renamed to `progaram::trap`, and the
+`program::abort` has been renamed to `program::trap`, and the
 "panic-handler-abort" feature has been renamed to "panic-handler-trap".
 
 # origin 0.22